moved ahab libs
